roberts comments
