+++ mirar/processors/photcal/zp_calculator/outlier_rejection_zp_calculator.py
@@ -1,48 +1,21 @@
 """
-Module for calculating zero points
+Module to calculate zero point using outlier rejection
 """
 import logging
 import warnings
-from typing import Callable
 
 import numpy as np
 from astropy.io.fits.verify import VerifyWarning
 from astropy.stats import sigma_clip, sigma_clipped_stats
 from astropy.table import Table
-from scipy.optimize import curve_fit
 
 from mirar.data import Image
-from mirar.processors.photcal import PhotometryCalculationError
+from mirar.processors.photcal.photcal_errors import PhotometryCalculationError
+from mirar.processors.photcal.zp_calculator.base_zp_calculator import (
+    BaseZeroPointCalculator,
+)
 
 logger = logging.getLogger(__name__)
-
-
-class BaseZeroPointCalculator:
-    """
-    Base class for actually calculating zero point calculators
-    """
-
-    def calculate_zeropoint(
-        self,
-        image: Image,
-        matched_ref_cat: Table,
-        matched_img_cat: Table,
-        colnames: list[str],
-    ) -> Image:
-        """
-        Function to calculate zero point from two catalogs and add the zeropoint
-        information to the image header
-        Args:
-            matched_ref_cat: Reference catalog table
-            matched_img_cat: Catalog of sources
-            image: Image object
-            colnames: List of column names from the image catalog to use for
-            calculating zero point. The reference catalog is assumed to have the
-            "magnitude" column, as it comes from mirar.catalog.base_catalog.BaseCatalog
-        Returns:
-            Image object with zero point information added to the header.
-        """
-        raise NotImplementedError
 
 
 class OutlierRejectionZPCalculator(BaseZeroPointCalculator):
@@ -116,60 +89,5 @@ class OutlierRejectionZPCalculator(BaseZeroPointCalculator):
                 image[f"ZP_{zpvals['diameter']}"] = zpvals["zp_mean"]
                 image[f"ZP_{zpvals['diameter']}_std"] = zpvals["zp_std"]
                 image[f"ZP_{zpvals['diameter']}_nstars"] = zpvals["nstars"]
-
-        return image
-
-
-def line_model(data, slope, intercept):
-    """linear model to hand to scipy curve_fit"""
-    return slope * data + intercept
-
-
-class ZPWithColorTermCalculator(BaseZeroPointCalculator):
-    """
-    Class to calculate zero point by including a color term. This models the data as
-
-    ref_mag - img_mag = ZP + C * (ref_color)
-
-    Attributes:
-        color_colnames_generator: function that takes an image as input and returns
-        a list containing two strings that are the column names of the reference catalog
-        to use for the color term. The first string is the bluer band, the second is the
-         redder band.
-    """
-
-    def __init__(
-        self,
-        color_colnames_generator: Callable[[Image], list[str, str]],
-    ):
-        self.color_colnames_generator = color_colnames_generator
-
-    def calculate_zeropoint(
-        self,
-        image: Image,
-        matched_ref_cat: Table,
-        matched_img_cat: Table,
-        colnames: list[str],
-    ) -> Image:
-        color_colnames = self.color_colnames_generator(image)
-        colors = matched_ref_cat[color_colnames[0]] - matched_ref_cat[color_colnames[1]]
-
-        for colname in colnames:
-            y = matched_img_cat[colname] - matched_ref_cat["magnitude"]
-            x = colors
-            y_err = np.sqrt(
-                matched_img_cat[colname.replace("MAG", "MAGERR")] ** 2
-                + matched_ref_cat["magnitude_err"] ** 2
-            )
-            popt, pcov = curve_fit(f=line_model, xdata=x, ydata=y, sigma=y_err)
-            color, zero_point = popt
-            color_err, zp_err = np.sqrt(np.diag(pcov))
-
-            aperture = colname.split("_")[-1]
-            image[f"ZP_{aperture}"] = zero_point
-            image[f"ZP_{aperture}_std"] = zp_err
-            image[f"ZP_{aperture}_nstars"] = len(matched_ref_cat)
-            image[f"C_{aperture}"] = color
-            image[f"C_{aperture}_std"] = color_err
 
         return image
